tests for Stellar Burger updated
